api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
supabase==1.4.0
python-dotenv==1.0.0
 
